# Refine auto spill for fine grained executor