--- a/dbms/src/Core/FineGrainedOperatorSpillContext.h
+++ b/dbms/src/Core/FineGrainedOperatorSpillContext.h
@@ -18,6 +18,16 @@
 
 namespace DB
 {
+/// FineGrainedOperatorSpillContext is a wrap for all the query spill contexts that belongs to a same executor
+/// with fine grained shuffle enable. When fine grained shuffle is enabled, executors like sort and aggregation
+/// will have n independent operators in TiFlash side, each of them have their own operator spill context. However
+/// these n operators are not really independent: them shared a same source operator(exchange receiver), if one of
+/// the operator stop consuming data from exchange receiver(for example, begin spill data), all the others will
+/// be stuck because the exchange receiver has a bounded queue for all the operator, any of the operator stop
+/// consuming will make the queue full, and no other data can be pushed to exchange receiver. If all the n operator
+/// is triggered to spill serially, it will affects the overall performance seriously.
+/// FineGrainedOperatorSpillContext is used to make sure that all the operators belongs to the same executor
+/// will be triggered to spill almost at the same time
 class FineGrainedOperatorSpillContext : public OperatorSpillContext
 {
 protected: