--- a/dbms/src/Flash/Coprocessor/DAGQueryBlockInterpreter.cpp
+++ b/dbms/src/Flash/Coprocessor/DAGQueryBlockInterpreter.cpp
@@ -522,8 +522,12 @@ void DAGQueryBlockInterpreter::executeAggregation(
     if (enable_fine_grained_shuffle)
     {
         std::shared_ptr<FineGrainedOperatorSpillContext> fine_grained_operator_spill_context;
-        if (context.getDAGContext() != nullptr && context.getDAGContext()->isInAutoSpillMode() && pipeline.hasMoreThanOneStream())
-            fine_grained_operator_spill_context = std::make_shared<FineGrainedOperatorSpillContext>(settings.max_bytes_before_external_group_by, "aggregation", log);
+        if (context.getDAGContext() != nullptr && context.getDAGContext()->isInAutoSpillMode()
+            && pipeline.hasMoreThanOneStream())
+            fine_grained_operator_spill_context = std::make_shared<FineGrainedOperatorSpillContext>(
+                settings.max_bytes_before_external_group_by,
+                "aggregation",
+                log);
         /// Go straight forward without merging phase when enable_fine_grained_shuffle
         pipeline.transform([&](auto & stream) {
             stream = std::make_shared<AggregatingBlockInputStream>(