### What problem does this PR solve?

Issue Number: ref #7738

Problem Summary:
When fine grained shuffle is enabled, due to implicit depentency of all the operator belongs to the same TiDB executor, auto spill will meet a big performance regression if auto spill treated all the operator spill contexts from the same TiDB executor independently. This pr add `FineGrainedOperatorSpillContext` and try to align the spill behavior for fine grained executor.
### What is changed and how it works?
- Add `FineGrainedOperatorSpillContext`
- For operator that comes from the same TiDB executor, auto spill will be triggered at TiDB executor level, instead of triggering each operator independently

Test on TPCH-100 dataset, with 2 TiFlash node
For fine grained agg, test sql is
```
select /*+ mpp_1phase_agg() */ l_orderkey, max(L_COMMENT), max(L_SHIPMODE), max(L_SHIPINSTRUCT), max(L_SHIPDATE), max(L_EXTENDEDPRICE) from lineitem group by l_orderkey having sum(l_quantity) > 314
``` 
per node memory usage limit to 12G:

| without this pr | with this pr| 
|-------------|--------------|
| ~45s            | ~33s |

per node memory usage limit to 7G:

| without this pr | with this pr| 
|-------------|--------------|
| ~37s            | ~30s |

For fine grained sort, test sql is
```
select * from (select L_ORDERKEY,L_COMMENT,L_SHIPINSTRUCT, row_number() over(partition by l_orderkey order by l_orderkey) as rn from lineitem) d where l_orderkey + rn < 50;
```
per node memory usage limit to 12G:

| without this pr | with this pr| 
|-------------|--------------|
| ~45s            | ~24s |

per node memory usage limit to 7G:

| without this pr | with this pr| 
|-------------|--------------|
| ~45s            | ~25s |
### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```