@SeaRise 2023-09-04
/hold
When you want to merge pr, comment `/hold cancel` @windtalker